refactor : parent allowance 폴더 위치 옮김
