+++ isol/server-parent/src/main/java/shinhan/server_parent/domain/allowance/repository/TemporalAllowanceRepositoryImpl.java
@@ -1,4 +1,4 @@
-package shinhan.server_common.domain.allowance.repository;
+package shinhan.server_parent.domain.allowance.repository;
 
 import jakarta.persistence.EntityManager;
 import org.springframework.beans.factory.annotation.Autowired;
